+++ src/App/Http/routes.php
@@ -44,6 +44,17 @@ Route::group(
 
 Route::group(
     [
+        'middleware' => 'Redooor\Redminportal\App\Http\Middleware\Authenticate',
+        'namespace' => 'Redooor\Redminportal\App\Http\API',
+        'prefix' => 'api'
+    ],
+    function () {
+        Route::controller('email', 'EmailApi');
+    }
+);
+
+Route::group(
+    [
         'namespace' => 'Redooor\Redminportal\App\Http\API',
         'prefix' => 'api'
     ],