+++ CHANGELOG.md
@@ -57,6 +57,10 @@ This update introduces some new tables and lots of changes.
     - HTML and Form helpers provide a shortcut to partial views.
     - Check [list of helpers on github Wiki page](https://github.com/redooor/redminportal/wiki/HTML-and-Form-Helpers)
 13. Added tag to Page and Post (issue #146)
+14. Added API for retrieving emails for authenticated admin users only
+    - /api/email: Unused, returns an empty JSON list
+    - /api/email/all: Get JSON list of emails
+    - Check [list of Private API on github Wiki page](https://github.com/redooor/redminportal/wiki/Private-API)
 
 ### Enhancements:
 1. Category model missing relationships with other models. (issue #140)
@@ -72,6 +76,8 @@ This update introduces some new tables and lots of changes.
 11. Introduced SorterController trait to reduce code for sortable pages.
 12. Introduced partial Blade template for sortable header.
 13. Overall Line count decreased from 23031 to 22805 with trait and partial template.
+14. Improve Usability: after delete action should go back to last visited page (issue #131)
+    - Order
 
 ### Bug fixes:
 1. User deletion: prevent user from deleting or deactivating own account while they are logged in. (issue #136)
@@ -82,6 +88,7 @@ This update introduces some new tables and lots of changes.
 6. Click menu overlay to close doesn't work on iPad. (issue #151)
 7. Fix Build error for PHP7 due to getID3 package. Switch to JamesHeinrich/getID3. (issue #152)
 8. Added delete image functionality to Bundles. (issue #149)
+9. Fixed Typeahead styling messed up for email suggestions. (issue #156)
 
 ## Version 0.3.2 and 0.2.2
 Code clean up, new features and UI improvements.